part-6: оптимизировали код, разделили на читаемую структуру, добавили комментарии и документацию.
