+++ README.md
@@ -1,11 +1,18 @@
-# Instapro
+# Instapro — учебный проект (рефакторинг по слоям)
 
-MVP социальной сети для обмена фотографиями
+## Быстрый старт
+1. Удалите старые файлы из корня: `api.js`, `helpers.js`, `index.js`, `routes.js`, папку `components/`.
+2. Скопируйте каталог `src/` из этого репозитория.
+3. В HTML подключите один файл:  
+   `<script type="module" src="./src/app/navigation.js"></script>`
+4. Запустите статический сервер (например `npx http-server`).
 
-## Первоначальная оценка
+## Структура
+- `src/api/*` — запросы к API: базовый клиент + модули по сущностям.
+- `src/app/*` — состояние, роутинг, навигация, глобальные события.
+- `src/components/*` — компоненты UI.
+- `src/lib/*` — вспомогалки (DOM/время/локальное хранилище).
 
-ХХХХ часов
-
-## Фактически затраченное время
-
-YYYY часов
+## Правила кода
+- Документируем экспортируемые функции через JSDoc.
+- Микро‑коммент «зачем» только там, где логика неочевидна.
