+++ components/pages/posts-page-component.js
@@ -1,25 +1,13 @@
-import {USER_POSTS_PAGE} from "../routes.js";
-import {renderHeaderComponent} from "./header-component.js";
-import {posts, goToPage, user} from "../index.js";
+import { USER_POSTS_PAGE } from "../../app/routes.js";
+import { renderHeaderComponent } from "../header/header-component.js";
+import { state } from "../../app/state.js";
+import { escapeHtml, timeAgo, ensureActionStyles } from "../../lib/dom.js";
 
-const escapeHtml = (s = "") =>
-    s.replace(/[&<>'"]/g, (c) => ({"&": "&amp;", "<": "&lt;", ">": "&gt;", "'": "&#39;", '"': "&quot;"}[c]));
-
-const timeAgo = (iso) => {
-    const d = new Date(iso);
-    if (isNaN(d)) return "";
-    const diffMs = Date.now() - d.getTime();
-    const m = Math.floor(diffMs / 60000);
-    if (m < 1) return "только что";
-    if (m < 60) return `${m} мин. назад`;
-    const h = Math.floor(m / 60);
-    if (h < 24) return `${h} ч. назад`;
-    const days = Math.floor(h / 24);
-    if (days < 7) return `${days} дн. назад`;
-    return d.toLocaleDateString();
-};
-
-// Нормализуем возможные варианты id
+/**
+ * Возвращает строковый id пользователя (пытается подобрать из разных полей).
+ * @param {any} u
+ * @returns {string}
+ */
 function getUserId(u) {
     if (!u) return "";
     if (u.id != null) return String(u.id);
@@ -29,53 +17,25 @@ function getUserId(u) {
     return "";
 }
 
-// Единый мини‑CSS для action‑кнопок (втыкаем 1 раз)
-function ensureActionStyles() {
-    if (document.getElementById("action-styles")) return;
-    const style = document.createElement("style");
-    style.id = "action-styles";
-    style.textContent = `
-    .danger-btn {
-      display:inline-flex;align-items:center;gap:6px;
-      padding:6px 10px;font-size:14px;line-height:1;
-      border-radius:8px;border:1px solid #e57373;background:#fff;color:#d32f2f;
-      cursor:pointer;transition:background .15s ease,border-color .15s ease,opacity .15s ease,transform .04s ease;
-    }
-    .danger-btn:hover { background:#ffebee;border-color:#ef5350; }
-    .danger-btn:active { transform:translateY(1px); }
-    .danger-btn:disabled { opacity:.6; cursor:not-allowed; }
-    .danger-btn .icon { width:16px;height:16px;display:inline-block }
-    .is-loading { pointer-events:none; }
-    .spinner { width:14px;height:14px;border:2px solid currentColor;border-right-color:transparent;border-radius:50%;
-               display:inline-block;animation:spin .8s linear infinite;vertical-align:-2px;margin-right:6px; }
-    @keyframes spin { to { transform: rotate(360deg) } }
-    .secondary-button {
-      display:inline-flex;align-items:center;gap:6px;padding:6px 10px;font-size:14px;line-height:1;border-radius:8px;
-      border:1px solid #d0d7de;background:#f6f8fa;color:#24292f;cursor:pointer;transition:background .15s ease,border-color .15s ease,transform .04s ease;
-    }
-    .secondary-button:hover { background:#eef2f6;border-color:#c0c7d0; }
-    .secondary-button:active { transform:translateY(1px); }
-    .secondary-button:disabled { opacity:.6; cursor:not-allowed; }
-  `;
-    document.head.appendChild(style);
-}
-
-export function renderPostsPageComponent({appEl}) {
+/**
+ * Страница ленты постов и постов пользователя.
+ * @param {{ appEl: HTMLElement }} params
+ */
+export function renderPostsPageComponent({ appEl }) {
     ensureActionStyles();
 
-    const currentUserId = getUserId(user);
+    const currentUserId = getUserId(state.user);
 
     const listHtml =
-        posts.length === 0
+        state.posts.length === 0
             ? `<li class="post post--empty">У пользователя пока нет постов.</li>`
-            : posts
-                .map((post) => {
-                    const isLiked = !!post.isLiked;
-                    const postUser = post.user || {};
-                    const postUserId = getUserId(postUser);
-                    const isOwner = currentUserId && postUserId && postUserId === currentUserId;
+            : state.posts.map((post) => {
+                const isLiked = !!post.isLiked;
+                const postUser = post.user || {};
+                const postUserId = getUserId(postUser);
+                const isOwner = currentUserId && postUserId && postUserId === currentUserId;
 
-                    return `
+                return `
   <li class="post" data-id="${post.id}">
     <div class="post-header" data-user-id="${postUserId}">
       <img src="${postUser.imageUrl}" class="post-header__user-image" alt="${escapeHtml(postUser.name || "")}">
@@ -128,8 +88,7 @@ export function renderPostsPageComponent({appEl}) {
     </p>
   </li>
 `;
-                })
-                .join("");
+            }).join("");
 
     const appHtml = `
     <div class="page-container">
@@ -146,28 +105,42 @@ export function renderPostsPageComponent({appEl}) {
         element: appEl.querySelector(".header-container"),
     });
 
+    // Переход на страницу постов пользователя
     appEl.querySelectorAll("[data-user-id]").forEach((userEl) => {
         userEl.addEventListener("click", () => {
             const uid = userEl.dataset.userId;
-            if (uid) goToPage(USER_POSTS_PAGE, {userId: uid});
+            if (!uid) return;
+            const ev = new CustomEvent("insta-navigate-user", { detail: { userId: uid }});
+            window.dispatchEvent(ev);
         });
     });
 
+    // Делегируем лайк глобальному обработчику, чтобы централизованно обновлять ленту
     appEl.querySelectorAll(".like-button").forEach((btn) => {
         btn.addEventListener("click", () => {
             const postId = btn.dataset.postId;
             const isLiked = btn.getAttribute("aria-pressed") === "true";
-            const ev = new CustomEvent("insta-like-click", {detail: {postId, isLiked}});
+            const ev = new CustomEvent("insta-like-click", { detail: { postId, isLiked } });
             window.dispatchEvent(ev);
         });
     });
 
+    // Удаление поста (только для владельца)
     appEl.querySelectorAll(".post-header__delete-button").forEach((btn) => {
         btn.addEventListener("click", (e) => {
             e.stopPropagation();
             const postId = btn.dataset.postId;
-            const ev = new CustomEvent("insta-delete-click", {detail: {postId}});
+            const ev = new CustomEvent("insta-delete-click", { detail: { postId } });
             window.dispatchEvent(ev);
         });
     });
+
+    // Навигация на USER_POSTS_PAGE
+    window.addEventListener("insta-navigate-user", (e) => {
+        const { userId } = e.detail || {};
+        if (!userId) return;
+        import("../../app/navigation.js").then(({ goToPage }) => {
+            goToPage(USER_POSTS_PAGE, { userId });
+        });
+    }, { once: true });
 }
