+++ docs/ARCHITECTURE.md
@@ -1,0 +1,10 @@
+# Архитектура
+
+## Слои
+- **API**: `src/api` — работа с сетью, маппинг ошибок. Не знает о DOM.
+- **APP**: `src/app` — состояние, навигация, события. Не рендерит разметку.
+- **UI Components**: `src/components` — чистый рендер HTML и биндинг событий.
+- **Lib**: `src/lib` — утилиты.
+
+## Поток данных
+UI → диспатчит кастомные события → `app/events` вызывает API → обновляет `app/state` → `app/navigation.renderApp()` перерисовывает.
