+++ docs/CODESTYLE.md
@@ -1,0 +1,8 @@
+# Code Style
+
+- Экспортируемые функции документируем JSDoc (1–2 строки «что делает», детали краёв, @param/@returns/@throws).
+- Константы страниц — в `app/routes.js`.
+- Состояние — в `app/state.js`, изменяем только через сеттеры.
+- Навигация и загрузка данных — в `app/navigation.js`.
+- Глобальные обработчики (лайк/удаление) — в `app/events.js`.
+- Утилиты DOM/времени/стилей — `lib/dom.js`.
