+++ index.html
@@ -13,5 +13,5 @@
     <div id="app"></div>
   </body>
 
-  <script type="module" src="./index.js"></script>
+  <script type="module" src="./app/navigation.js"></script>
 </html>
