--- index.js
@@ -1,193 +1,0 @@
-import { addPost, deletePost, getPosts, getUserPosts, toggleLike } from "./api.js";
-import { renderAddPostPageComponent } from "./components/add-post-page-component.js";
-import { renderAuthPageComponent } from "./components/auth-page-component.js";
-import { ADD_POSTS_PAGE, AUTH_PAGE, LOADING_PAGE, POSTS_PAGE, USER_POSTS_PAGE } from "./routes.js";
-import { renderPostsPageComponent } from "./components/posts-page-component.js";
-import { renderLoadingPageComponent } from "./components/loading-page-component.js";
-import { getUserFromLocalStorage, removeUserFromLocalStorage, saveUserToLocalStorage } from "./helpers.js";
-
-export let user = getUserFromLocalStorage();
-export let page = null;
-export let posts = [];
-
-let currentUserId = null;
-
-const getToken = () => {
-    const token = user ? `Bearer ${user.token}` : undefined;
-    return token;
-};
-
-export const logout = () => {
-    user = null;
-    removeUserFromLocalStorage();
-    goToPage(POSTS_PAGE);
-};
-
-/**
- * Включает страницу приложения
- */
-export const goToPage = (newPage, data) => {
-    if (
-        [POSTS_PAGE, AUTH_PAGE, ADD_POSTS_PAGE, USER_POSTS_PAGE, LOADING_PAGE].includes(
-            newPage
-        )
-    ) {
-        if (newPage === ADD_POSTS_PAGE) {
-            page = user ? ADD_POSTS_PAGE : AUTH_PAGE;
-            return renderApp();
-        }
-
-        if (newPage === USER_POSTS_PAGE) {
-            page = LOADING_PAGE;
-            renderApp();
-
-            currentUserId = data?.userId ?? null;
-
-            return getUserPosts({ userId: currentUserId, token: getToken() })
-                .then((newPosts) => {
-                    page = USER_POSTS_PAGE;
-                    posts = newPosts;
-                    renderApp();
-                })
-                .catch((error) => {
-                    console.error(error);
-                    goToPage(POSTS_PAGE);
-                });
-        }
-
-        if (newPage === POSTS_PAGE) {
-            page = LOADING_PAGE;
-            renderApp();
-
-            return getPosts({ token: getToken() })
-                .then((newPosts) => {
-                    posts = newPosts;
-                    page = POSTS_PAGE;
-                    renderApp();
-                })
-                .catch((error) => {
-                    console.error(error);
-                    posts = [];
-                    page = POSTS_PAGE;
-                    renderApp();
-                });
-        }
-
-        page = newPage;
-        renderApp();
-
-        return;
-    }
-
-    throw new Error("страницы не существует");
-};
-
-const renderApp = () => {
-    const appEl = document.getElementById("app");
-    if (page === LOADING_PAGE) {
-        return renderLoadingPageComponent({
-            appEl,
-            user,
-            goToPage,
-        });
-    }
-
-    if (page === AUTH_PAGE) {
-        return renderAuthPageComponent({
-            appEl,
-            setUser: (newUser) => {
-                user = newUser;
-                saveUserToLocalStorage(user);
-                goToPage(POSTS_PAGE);
-            },
-            user,
-            goToPage,
-        });
-    }
-
-    if (page === ADD_POSTS_PAGE) {
-        return renderAddPostPageComponent({
-            appEl,
-            async onAddPostClick({ description, imageUrl }) {
-                try {
-                    await addPost({ description, imageUrl, token: getToken() });
-                    goToPage(POSTS_PAGE);
-                } catch (e) {
-                    alert(e.message || "Не удалось добавить пост");
-                }
-            },
-        });
-    }
-
-    if (page === POSTS_PAGE) {
-        return renderPostsPageComponent({
-            appEl,
-        });
-    }
-
-    if (page === USER_POSTS_PAGE) {
-        return renderPostsPageComponent({
-            appEl,
-        });
-    }
-};
-
-window.addEventListener("insta-like-click", async (e) => {
-    const { postId, isLiked } = e.detail || {};
-    if (!postId) return;
-
-    try {
-        await toggleLike({ postId, isLiked, token: getToken() });
-        if (page === USER_POSTS_PAGE && currentUserId) {
-            await getUserPosts({ userId: currentUserId, token: getToken() }).then(
-                (newPosts) => {
-                    posts = newPosts;
-                    renderApp();
-                }
-            );
-        } else {
-            await getPosts({ token: getToken() }).then((newPosts) => {
-                posts = newPosts;
-                renderApp();
-            });
-        }
-    } catch (err) {
-        alert(err.message || "Не удалось обновить лайк");
-    }
-});
-
-window.addEventListener("insta-delete-click", async (e) => {
-    const { postId } = e.detail || {};
-    if (!postId) return;
-    if (!user || !user.token) {
-        alert("Войдите, чтобы удалить пост");
-        return;
-    }
-    if (!confirm("Удалить пост без возможности восстановления?")) return;
-
-    // Блокируем кнопку и показываем спиннер
-    const btn = document.querySelector(`.post-header__delete-button[data-post-id="${postId}"]`);
-    const originalHtml = btn?.innerHTML;
-    if (btn) {
-        btn.disabled = true;
-        btn.classList.add("is-loading");
-        btn.innerHTML = `<span class="spinner" aria-hidden="true"></span>Удаляю…`;
-    }
-
-    try {
-        await deletePost({ postId, token: getToken() });
-        posts = (page === USER_POSTS_PAGE && currentUserId)
-            ? await getUserPosts({ userId: currentUserId, token: getToken() })
-            : await getPosts({ token: getToken() });
-        renderApp();
-    } catch (err) {
-        if (btn) {
-            btn.disabled = false;
-            btn.classList.remove("is-loading");
-            btn.innerHTML = originalHtml ?? "Удалить";
-        }
-        alert(err.message || "Не удалось удалить пост");
-    }
-});
-
-goToPage(POSTS_PAGE);
